requirements.txt
web3
python-decouple

# Test
web3[tester]
py-solc-x
